Refactor: Consolidate CLI commands and move adapters to integrations
