--- src/rldk/adapters/__init__.py
@@ -1,15 +1,0 @@
-"""Adapters for different training log formats."""
-
-from .base import BaseAdapter
-from .trl import TRLAdapter
-from .openrlhf import OpenRLHFAdapter
-from .wandb import WandBAdapter
-from .custom_jsonl import CustomJSONLAdapter
-
-__all__ = [
-    "BaseAdapter",
-    "TRLAdapter",
-    "OpenRLHFAdapter",
-    "WandBAdapter",
-    "CustomJSONLAdapter",
-]
