+++ src/rldk/integrations/__init__.py
@@ -1,3 +1,13 @@
 """RLDK integrations package for third-party libraries."""
 
+from .trl import TRLAdapter
+from .openrlhf import OpenRLHFAdapter
+from .wandb import WandBAdapter
+
 __version__ = "0.1.0"
+
+__all__ = [
+    "TRLAdapter",
+    "OpenRLHFAdapter", 
+    "WandBAdapter",
+]
