+++ src/rldk/integrations/wandb/__init__.py
@@ -1,0 +1,5 @@
+"""Weights & Biases integration for RL Debug Kit."""
+
+from .adapter import WandBAdapter
+
+__all__ = ["WandBAdapter"]
